--- a/_examples/chat_json/main.go
+++ b/_examples/chat_json/main.go
@@ -90,6 +90,11 @@ func main() {
 
 	node, _ := centrifuge.New(cfg)
 
+	engine, _ := centrifuge.NewMemoryEngine(node, centrifuge.MemoryEngineConfig{
+		SequenceTTL: 120 * time.Second,
+	})
+	node.SetEngine(engine)
+
 	node.On().ClientConnecting(func(ctx context.Context, t centrifuge.TransportInfo, e centrifuge.ConnectEvent) centrifuge.ConnectReply {
 		return centrifuge.ConnectReply{
 			Data: centrifuge.Raw(`{}`),