--- a/_examples/chat_json/main.go
+++ b/_examples/chat_json/main.go
@@ -71,7 +71,7 @@ func main() {
 	cfg.UserSubscribeToPersonal = true
 
 	cfg.Namespaces = []centrifuge.ChannelNamespace{
-		centrifuge.ChannelNamespace{
+		{
 			Name: "chat",
 			ChannelOptions: centrifuge.ChannelOptions{
 				Publish:         true,