--- a/_examples/go.mod
+++ b/_examples/go.mod
@@ -6,6 +6,7 @@ replace github.com/centrifugal/centrifuge => ../
 
 require (
 	github.com/centrifugal/centrifuge v0.7.0
+	github.com/centrifugal/protocol v0.2.0
 	github.com/dchest/uniuri v0.0.0-20200228104902-7aecb25e1fe5
 	github.com/gin-contrib/sessions v0.0.3
 	github.com/gin-gonic/gin v1.5.0