--- a/_examples/worms/main.go
+++ b/_examples/worms/main.go
@@ -60,7 +60,7 @@ func main() {
 		client.On().Message(func(e centrifuge.MessageEvent) centrifuge.MessageReply {
 			var ev event
 			_ = json.Unmarshal(e.Data, &ev)
-			node.Publish("moving", []byte(ev.Payload))
+			_ = node.Publish("moving", []byte(ev.Payload))
 			return centrifuge.MessageReply{}
 		})