This pull request replaces a data structure inside Memory engine to in-memory `Stream` data structure which is a combination of linked list and hash map for quick access from certain position.

Also there are some refactoring to get rid of some type aliases, many linter fixes. 

`RecoveryPosition` type renamed to `StreamPosition`. I also have some thoughts on next steps in library development: remove all exported aliases if possible, investigate Redis Streams as option for Redis Engine, introduce new history API to paginate over stream. But this all may come later.

Benchmarks:

```
benchmark                                                             old ns/op     new ns/op     delta
BenchmarkMemoryEnginePublish_WithHistory_SingleChannel-4              1086          446           -58.93%
BenchmarkMemoryEnginePublish_WithHistory_SingleChannel_Parallel-4     1495          812           -45.69%
BenchmarkMemoryEngineHistory_SingleChannel-4                          257           245           -4.67%
BenchmarkMemoryEngineHistory_SingleChannel_Parallel-4                 214           334           +56.07%
BenchmarkMemoryEngineRecover_SingleChannel_Parallel-4                 1011          413           -59.15%
BenchmarkWebsocketHandler/JSON-4                                      35337         34586         -2.13%
BenchmarkWebsocketHandler/Protobuf-4                                  29932         31238         +4.36%
BenchmarkNodePublishWithNoopEngine-4                                  163           156           -4.29%
BenchmarkBroadcastMemoryEngine/JSON_1_subscribers-4                   5470          5705          +4.30%
BenchmarkBroadcastMemoryEngine/Protobuf_1_subscribers-4               5604          5961          +6.37%
BenchmarkBroadcastMemoryEngine/JSON_10000_subscribers-4               10125995      10338205      +2.10%
BenchmarkBroadcastMemoryEngine/Protobuf_10000_subscribers-4           10119306      10238695      +1.18%

benchmark                                                             old allocs     new allocs     delta
BenchmarkMemoryEnginePublish_WithHistory_SingleChannel-4              3              2              -33.33%
BenchmarkMemoryEnginePublish_WithHistory_SingleChannel_Parallel-4     4              3              -25.00%
BenchmarkMemoryEngineHistory_SingleChannel-4                          1              2              +100.00%
BenchmarkMemoryEngineHistory_SingleChannel_Parallel-4                 1              2              +100.00%
BenchmarkMemoryEngineRecover_SingleChannel_Parallel-4                 1              2              +100.00%
BenchmarkWebsocketHandler/JSON-4                                      14             14             +0.00%
BenchmarkWebsocketHandler/Protobuf-4                                  14             14             +0.00%
BenchmarkNodePublishWithNoopEngine-4                                  3              3              +0.00%
BenchmarkBroadcastMemoryEngine/JSON_1_subscribers-4                   11             11             +0.00%
BenchmarkBroadcastMemoryEngine/Protobuf_1_subscribers-4               11             11             +0.00%
BenchmarkBroadcastMemoryEngine/JSON_10000_subscribers-4               13             13             +0.00%
BenchmarkBroadcastMemoryEngine/Protobuf_10000_subscribers-4           14             15             +7.14%

benchmark                                                             old bytes     new bytes     delta
BenchmarkMemoryEnginePublish_WithHistory_SingleChannel-4              978           87            -91.10%
BenchmarkMemoryEnginePublish_WithHistory_SingleChannel_Parallel-4     1045          151           -85.55%
BenchmarkMemoryEngineHistory_SingleChannel-4                          32            160           +400.00%
BenchmarkMemoryEngineHistory_SingleChannel_Parallel-4                 32            160           +400.00%
BenchmarkMemoryEngineRecover_SingleChannel_Parallel-4                 896           176           -80.36%
BenchmarkWebsocketHandler/JSON-4                                      2569          2569          +0.00%
BenchmarkWebsocketHandler/Protobuf-4                                  2473          2473          +0.00%
BenchmarkNodePublishWithNoopEngine-4                                  97            97            +0.00%
BenchmarkBroadcastMemoryEngine/JSON_1_subscribers-4                   513           513           +0.00%
BenchmarkBroadcastMemoryEngine/Protobuf_1_subscribers-4               513           513           +0.00%
BenchmarkBroadcastMemoryEngine/JSON_10000_subscribers-4               699           684           -2.15%
BenchmarkBroadcastMemoryEngine/Protobuf_10000_subscribers-4           711           871           +22.50%
```